Shorten MaxMessagesPerCommit
This is required by BoundedVec, as its implementation for Decode expects
the bound to be a u32.

+++ parachain/pallets/basic-channel/src/outbound/mod.rs
@@ -74,7 +74,7 @@ pub mod pallet {
 
 		/// Max number of messages per commitment
 		#[pallet::constant]
-		type MaxMessagesPerCommit: Get<u64>;
+		type MaxMessagesPerCommit: Get<u32>;
 
 		type SetPrincipalOrigin: EnsureOrigin<Self::Origin>;
